Add regex content and update project structure.
