+++ README.txt
@@ -39,3 +39,13 @@ File content can also be random::
  print len(sfs.open('random/128KB+1').read())
  131073
 
+
+File content can be generated by restricted regex pattern by adding a directory::
+
+ sfs.add_regex_dir("regex1","a(bcd)*e{4}[a-z,0,3]*")
+ print len(sfs.open('regex1/128KB').read())
+ 131072
+ print len(sfs.open('regex1/128KB-1').read())
+ 131071
+ print len(sfs.open('regex1/128KB+1').read())
+ 131073
